Add failing increment_after_colon test
It's legal in Python to put a statement on the same line after the
colon of statements like `for`, `if`, `while`, etc. In this
circumstance, the experimental increment operator fails to parse. I
have no fix for this issue, but here is a simple test case.
Unfortunately, the test can't actually be run, since the syntax it
would check is not legal.